[BugFix] ios/android device can set param of vertical/horizontal distance correctly.
    [id]:102079

    [Type]
        BugFix

    [Products]
        HiFly

    [Requirement]
        N/A

    [Solution Description]
        N/A

    [Other Info]
        N/A

+++ src/modules/navigator/geofence.cpp
@@ -144,8 +144,8 @@ bool Geofence::inside(double lat, double lon, float altitude, struct geofence_re
 	result.geofence_hor_violated = false;
 
 	if (_param_geofence_mode.get() >= GEOFENCE_MAX_DISTANCES_ONLY) {
-		int32_t max_horizontal_distance = _param_max_hor_distance.get();
-		int32_t max_vertical_distance = _param_max_ver_distance.get();
+		float max_horizontal_distance = _param_max_hor_distance.get();
+		float max_vertical_distance = _param_max_ver_distance.get();
 
 		if (max_horizontal_distance > 0 || max_vertical_distance > 0) {
 			if (_home_pos_set) {
